chore(refactoring): refactoring code for health tracker processor
