+++ HealthTrackerProcessor/Class/DBContextConfiguration.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HealthTrackerProcessor.Class
+{
+    public interface IDBContextConfiguration
+    {
+        string DatabaseName { get; }
+        string ConnectionString { get; }
+    }
+}
