+++ HealthTrackerProcessor/Class/User.cs
@@ -1,0 +1,24 @@
+﻿using Castle.MicroKernel;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using static HealthTrackerProcessor.Enum.HealthTrackerProcessorEnum;
+
+namespace HealthTrackerProcessor.Class
+{
+    public interface IUser
+    {
+        public string ID { get; set; }
+        public string Username { get; set; }
+        public Gender Gender { get; set; }
+    }
+    public class User : IUser
+    {
+        public string ID { get; set; } = Guid.NewGuid().ToString();
+        public string Username { get; set; }
+        public double Height { get; set; }
+        public Gender Gender { get; set; }
+
+    }
+}
