+++ HealthTrackerProcessor/Constant/HealthTrackerProcessorConstant.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HealthTrackerProcessor.Constant
+{
+    public class HealthTrackerProcessorConstant
+    {
+        public const string SQLiteDBConnectionStr = "Data Source=Data.db";
+        public const string PostgresDBConnectionStr = "Host=172.17.0.2;Username=postgres;Port=5432;Password=postgres";
+    }
+}
