+++ HealthTrackerProcessor/Enum/HealthTrackerProcessorEnum.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HealthTrackerProcessor.Enum
+{
+    public static class HealthTrackerProcessorEnum
+    {
+        public enum ProjectType
+        {
+            Sandbox = 0,
+            Application = 1,
+        }
+        public enum Gender
+        {
+            Female = 0,
+            Male = 1,
+        }
+    }
+}
