+++ HealthTrackerProcessor/Handler/RawDataHandler.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HealthTrackerProcessor.Handler
+{
+    public class RawDataHandler
+    {
+
+    }
+}
