+++ HealthTrackerProcessorCore/Class/AjaxResponse.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace HealthTrackerProcessorCore.Class
+{
+    public class AjaxResponse
+    {
+        public bool Succeed { get; set; }
+        public string Status { get; set; }
+        public string Message { get; set; }
+        public object Data { get; set; }
+    }
+}
