+++ HealthTrackerProcessorCore/DatabaseUtil.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore;
+
+namespace HealthTrackerProcessorCore
+{
+    public class DatabaseUtil
+    {
+    }
+}
