+++ HealthTrackerProcessorCore/HealthTrackerProcessorCore.csproj
