--- MyWebApp.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30330.147
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyWebApp", "MyWebApp\MyWebApp.csproj", "{4C5C2418-2486-491E-8767-4119F8AC3658}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyWebAppCore", "MyWebAppCore\MyWebAppCore.csproj", "{B0CE1B6E-DE3C-4175-8784-638AE7117C32}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4C5C2418-2486-491E-8767-4119F8AC3658}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C5C2418-2486-491E-8767-4119F8AC3658}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C5C2418-2486-491E-8767-4119F8AC3658}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C5C2418-2486-491E-8767-4119F8AC3658}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0CE1B6E-DE3C-4175-8784-638AE7117C32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0CE1B6E-DE3C-4175-8784-638AE7117C32}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0CE1B6E-DE3C-4175-8784-638AE7117C32}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0CE1B6E-DE3C-4175-8784-638AE7117C32}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E2FCEF01-8693-4EC6-8433-989CBAB4E401}
-	EndGlobalSection
-EndGlobal
