--- MyWebApp/Class/DBContextConfiguration.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace MyWebApp.Class
-{
-    public interface IDBContextConfiguration
-    {
-        string DatabaseName { get; }
-        string ConnectionString { get; }
-    }
-}
