--- MyWebApp/Enum/MyWebAppEnum.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace MyWebApp.Enum
-{
-    public static class MyWebAppEnum
-    {
-        public enum ProjectType
-        {
-            Sandbox = 0,
-            Application = 1,
-        }
-        public enum Gender
-        {
-            Female = 0,
-            Male = 1,
-        }
-    }
-}
