--- MyWebApp/Handler/RawDataHandler.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace MyWebApp.Handler
-{
-    public class RawDataHandler
-    {
-
-    }
-}
