--- MyWebAppCore/DatabaseUtil.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using Microsoft.EntityFrameworkCore;
-
-namespace MyWebAppCore
-{
-    public class DatabaseUtil
-    {
-    }
-}
